// backend/src/common/enums.ts
export enum TaskStatus {
  TODO = 'TODO',
  DOING = 'DOING',
  COMPLETED = 'COMPLETED',
  ON_HOLD = 'ON_HOLD',
}

export enum Priority {
  NONE = 'NONE',
  URGENT = 'URGENT',
  HIGH = 'HIGH',
  MEDIUM = 'MEDIUM',
  LOW = 'LOW',
}

export enum AccentColor {
  AMBER = 'AMBER',
  BLUE = 'BLUE',
  PINK = 'PINK',
  ROSE = 'ROSE',
  EMERALD = 'EMERALD',
  BLACK = 'BLACK',
}

export enum Theme {
  LIGHT = 'light',
  DARK = 'dark',
}